+++ src/AssetNXT.Client/AssetNXT.Client.njsproj
@@ -90,11 +90,11 @@
     <Content Include="src\components\assets\telemetric\TelemetricTabs.js">
       <SubType>Code</SubType>
     </Content>
-    <Content Include="src\components\constrains\createConstrains.js" />
-    <Content Include="src\components\constrains\deleteConstrains.js" />
-    <Content Include="src\components\constrains\editConstrains.js" />
-    <Content Include="src\pages\ConstrainPage.css" />
-    <Content Include="src\pages\ConstrainPage.js" />
+    <Content Include="src\components\constrains\actions\createConstrains.js" />
+    <Content Include="src\components\constrains\actions\deleteConstrains.js" />
+    <Content Include="src\components\constrains\actions\editConstrains.js" />
+    <Content Include="src\components\constrains\ConstrainService.css" />
+    <Content Include="src\components\constrains\ConstrainService.js" />
     <Content Include="src\components\notifications\NotificationList.css">
       <SubType>Code</SubType>
     </Content>
@@ -118,6 +118,9 @@
     <Content Include="src\components\search\Searchbar.css" />
     <Content Include="src\components\search\Searchbar.js" />
     <Content Include="src\index.js" />
+    <Content Include="src\pages\ConstrainPage.js">
+      <SubType>Code</SubType>
+    </Content>
     <Content Include="src\pages\StationPage.css">
       <SubType>Code</SubType>
     </Content>
@@ -141,6 +144,7 @@
     <Folder Include="src\components\assets\telemetric\" />
     <Folder Include="src\components\assets\telemetric\graphs\" />
     <Folder Include="src\components\constrains\" />
+    <Folder Include="src\components\constrains\actions\" />
     <Folder Include="src\components\notifications\" />
     <Folder Include="src\components\search\" />
     <Folder Include="src\pages\" />
