+++ src/AssetNXT.Client/src/components/constrains/actions/deleteConstrains.js
